Move memory-related functions to memory.c
Sort includes in some sources.
