+++ src/parser.c
@@ -23,6 +23,7 @@
 
 #include "include/expr.h"
 #include "include/util.h"
+#include "include/memory.h"
 #include "include/lexer.h"
 #include "include/parser.h"
