+++ src/util.c
@@ -16,7 +16,6 @@
  * SL. If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include <errno.h>
 #include <stdarg.h>
 #include <stdbool.h>
 #include <stdio.h>
@@ -25,32 +24,8 @@
 #include <regex.h>
 
 #include "include/util.h"
+#include "include/memory.h"
 #include "include/error.h"
-
-void* sl_safe_malloc(size_t size) {
-    void* result = malloc(size);
-    if (result == NULL)
-        SL_FATAL("Failed to allocate %zu bytes: %s (%d).", size,
-                 strerror(errno), errno);
-    return result;
-}
-
-void* sl_safe_calloc(size_t nmemb, size_t size) {
-    void* result = calloc(nmemb, size);
-    if (result == NULL)
-        SL_FATAL("Failed to allocate %zu elements of %zu bytes each: %s (%d).",
-                 nmemb, size, strerror(errno), errno);
-    return result;
-}
-
-char* sl_safe_strdup(const char* s) {
-    char* result = strdup(s);
-    if (result == NULL)
-        SL_FATAL("Failed to copy string: %s (%d).", strerror(errno), errno);
-    return result;
-}
-
-/*----------------------------------------------------------------------------*/
 
 /* clang-format off */
 char escaped2byte(char escaped) {
